c99 removed
